+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix encoded headers
 - fix attainable contents fetching
 - embedding inline images 
+- loading parses attachments
 
 ## 1.2 (2020-06-18)
 - fix: smtplib.SMTP_SSL handshake operation timed out caught